Generate a Final Labeled Dataframe

This function processes an input dataset by combining and selecting relevant columns (identifiers with highest confidence levels)

final_label_dataframe(data_input, details = FALSE, original_name = "")

Arguments

data_input

A data frame containing input data with columns for accession and name information.

details

Logical, indicating whether to include all details in the output. If `TRUE`, the entire dataframe is returned with renamed columns for clarity. If `FALSE`, only selected columns are returned.

original_name

A character string representing the column name in the input data that identifies the original study.

Value

A data frame with the final processed results: - If `details = TRUE`, all processed columns are returned with renamed columns for clarity. - If `details = FALSE`, a subset with columns for the original study name, HMDB label, HMDB identifier, and confidence level is returned.

Details

The function performs the following steps: 1. Identifies columns starting with `"accession_in_reference"` and `"name_in_reference"`. 2. Creates new columns `HMDB_id_list` and `HMDB_name_list` by combining unique, non-NA values from the identified columns. 3. Extracts the highest confidence level (based on the number of stars) for `HMDB_id` and `HMDB_label` using the `extract_highest_star` function. 4. Calculates the number of stars (`Num_Stars`) for the label and maps star counts to corresponding confidence levels. 5. Depending on the `details` parameter, either the full processed dataframe or a subset with selected columns is returned.